[add] 프로젝트 생성시 알람오는 기능 구현
