+++ package.json
@@ -22,6 +22,7 @@
     "@tiptap/pm": "^2.1.15",
     "@tiptap/react": "^2.1.15",
     "@tiptap/starter-kit": "^2.1.15",
+    "cheerio": "^1.0.0-rc.12",
     "dayjs": "^1.11.10",
     "framer-motion": "^10.17.9",
     "next": "14.0.4",
