+++ src/app/(providers)/(default)/_components/page.tsx
@@ -1,0 +1,13 @@
+import React from "react"
+import Header from "./Header"
+const Page = () => {
+  return (
+    <div>
+      <Alarm>
+        <Header />
+      </Alarm>
+    </div>
+  )
+}
+
+export default Page
